requirements.txt
"path": "C:\\Users\\siraw\\.cache\\kagglehub\\datasets\\olistbr\\brazilian-ecommerce\\versions\\2"
"path": "/Users/robot/.cache/kagglehub/datasets/olistbr/brazilian-ecommerce/versions/2"
python: 3.9
anaconda env: dbt-env
kagglehub>=0.4.1
pandas
google-cloud-bigquery
pyarrow